requirements.txt
PyQt6>=6.6.0
oracledb>=2.0.0
PyYAML>=6.0
